+++ src/sentry/android/android_string_names.h
@@ -5,7 +5,7 @@
 
 using namespace godot;
 
-namespace sentry {
+namespace sentry::android {
 
 /**
  * Stores StringName constants for Android SDK implementation.
@@ -74,8 +74,8 @@ public:
 	StringName eventAddException;
 };
 
-} // namespace sentry
+} //namespace sentry::android
 
-#define ANDROID_SN(m_arg) sentry::AndroidStringNames::get_singleton()->m_arg
+#define ANDROID_SN(m_arg) sentry::android::AndroidStringNames::get_singleton()->m_arg
 
 #endif // SENTRY_ANDROID_STRING_NAMES_H